requirements.txt
flask
flask-socketio
python-dotenv
eventlet
flask-cors
alpaca-py